Add files via upload
Adding points to a list and going through them causes a pystack exhausted error for the CG-50. I don't know how to fix it right now, so now it'll output two scripts: code1.py is the one with the points in a list. It is significantly lighter than code2.py in filesize, but could get you that error. code2.py is (relatively) huge, but won't give you any issues if you're at 16kb for the 9750GIII or 132 for the CG50.